+++ KSP2Runtime/KSPOAB/KSPOABModule.ObjectAssemblyCommand.cs
@@ -1,0 +1,14 @@
+﻿using KontrolSystem.KSP.Runtime.KSPVessel;
+using KontrolSystem.TO2.Binding;
+using KSP.Modules;
+using KSP.OAB;
+
+namespace KontrolSystem.KSP.Runtime.KSPOAB;
+
+public partial class KSPOABModule {
+    [KSClass("ObjectAssemblyCommand")]
+    public class ObjectAssemblyCommandAdapter : BaseCommandAdapter<ObjectAssemblyPartAdapter, IObjectAssemblyPart> {
+        public ObjectAssemblyCommandAdapter(ObjectAssemblyPartAdapter part, Data_Command dataCommand) : base(part, dataCommand) {
+        }
+    }
+}